users_service/requirements.txt
fastapi
uvicorn[standard]
email-validator
sqlalchemy
psycopg2-binary
pytest
flake8
black
isort
requests
python-dotenv
pytest
pytest-asyncio 
httpx 
coverage
alembic